Moved classes to specific folders and improved them a bit
